feat: resolve issues #314, #315, #320, #321 — batch distribute, router decoupling, frontend decomposition, event listener
## Changes

### Contracts (#315: batch_distribute)
- Added batch_distribute function for distributing to multiple projects in one call
- Added 3 new contract tests for batch distribution scenarios

### Backend (#320: Router decoupling, #314: Event listener)
- Decoupled monolithic splits router into focused service modules
- Created EventListenerService for polling Soroban contract events
- Fixed duplicate logger import in stellar.ts
- Fixed GetEventsRequest discriminated union type handling

### Frontend (#321: Decompose split-app.tsx)
- Decomposed 1400-line split-app.tsx into focused components:
  DashboardView, CreateSplitWizard, ManageSplitView, ProjectsList
- Added camelCase mapping layer in api-client.ts for type safety
- Fixed wallet network parsing for Freighter SDK compatibility
- Updated test suite for new component structure

### Type Safety
- Unified WalletState and SplitProject interfaces in stellar.ts
- Added mapProjectToCamelCase compatibility layer
- Fixed StellarWalletsKit.setNetwork type cast

Closes #314
Closes #315
Closes #320
Closes #321
