+++ contracts/interface/splitnaira.contract-interface.json
@@ -2,7 +2,7 @@
   "schema": "https://splitnaira.dev/schemas/contract-interface.v1.json",
   "schemaVersion": 1,
   "generatedBy": "contracts/scripts/generate-interface.mjs",
-  "sourceHash": "c2087d1920bb3c4b6cbb3134a36e15a318791d2664c5af8bd6cf66b083a3e202",
+  "sourceHash": "e96ffc07d11d8860d9d11f88feb5d1e447a10d1be558269d6bb6bb5c3d19ba0b",
   "sources": [
     "contracts/Cargo.toml",
     "contracts/lib.rs",
@@ -169,6 +169,17 @@
         {
           "name": "project_id",
           "type": "Symbol"
+        }
+      ],
+      "returnType": "Result<(), SplitError>",
+      "mutability": "write"
+    },
+    {
+      "name": "batch_distribute",
+      "args": [
+        {
+          "name": "project_ids",
+          "type": "Vec<Symbol>"
         }
       ],
       "returnType": "Result<(), SplitError>",
