+++ frontend/src/lib/wallet.ts
@@ -44,13 +44,16 @@ function parseNetwork(network: string): string {
 export async function getWalletState(): Promise<WalletState> {
   try {
     const { address } = await StellarWalletsKit.getAddress();
-    const network = StellarWalletsKit.getNetwork
+    const rawNetwork = StellarWalletsKit.getNetwork
       ? await StellarWalletsKit.getNetwork()
       : null;
+    const resolvedNetwork = typeof rawNetwork === "object" && rawNetwork !== null
+      ? (rawNetwork as any).network
+      : rawNetwork;
     return {
       connected: true,
       address: address ?? null,
-      network: network ?? null,
+      network: typeof resolvedNetwork === "string" ? parseNetwork(resolvedNetwork) : null,
     };
   } catch {
     return { connected: false, address: null, network: null };
@@ -59,15 +62,18 @@ export async function getWalletState(): Promise<WalletState> {
 
 export async function connectWallet(network?: string): Promise<WalletState> {
   const targetNetwork = network ?? "TESTNET";
-  StellarWalletsKit.setNetwork(targetNetwork);
+  StellarWalletsKit.setNetwork(targetNetwork as any);
   const { address } = await StellarWalletsKit.authModal();
-  const walletNetwork = StellarWalletsKit.getNetwork
+  const rawNetwork = StellarWalletsKit.getNetwork
     ? await StellarWalletsKit.getNetwork()
     : targetNetwork;
+  const resolvedNetwork = typeof rawNetwork === "object" && rawNetwork !== null
+    ? (rawNetwork as any).network
+    : rawNetwork;
   return {
     connected: true,
     address: address ?? null,
-    network: walletNetwork ?? targetNetwork,
+    network: typeof resolvedNetwork === "string" ? parseNetwork(resolvedNetwork) : targetNetwork,
   };
 }
